Pushing my code

+++ trust_region_irl/algorithms/data_utils.py
@@ -46,13 +46,15 @@ def prepare_expert_data(data_path, cutoff=1):
     expert_files = np.load(data_path)
 
     def _flatten_feature_array(x):
-        x = np.asarray(x)
+        # Cast to float32 to match expert_data_spec (io_callback requires an
+        # exact dtype match; some datasets store e.g. bool/float64 fields).
+        x = np.asarray(x, dtype=np.float32)
         if x.ndim <= 2:
             return x
         return x.reshape(-1, x.shape[-1])   # (30,1000,d) -> (30000,d)
 
     def _flatten_scalar_array(x):
-        return np.asarray(x).reshape(-1)     # (30,1000) -> (30000,)
+        return np.asarray(x, dtype=np.float32).reshape(-1)     # (30,1000) -> (30000,)
 
     # load and normalize expert training data
     states = _flatten_feature_array(expert_files["states"])
